final_project_revised/django/requirements.txt
Django>=4.2
google-generativeai>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
# run pip install -r requirements.txt
# to install the dependencies
# and run pip install --upgrade google-generativeai